backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
httpx==0.28.1
python-dotenv==1.1.0
pytest==8.3.5
pytest-asyncio==0.25.3
fpdf2==2.8.3
